+++ examples/configs/src/index.js
@@ -10,7 +10,7 @@ import { spraggins2020, neumann2020 } from './view-configs/spraggins.js';
 import { satija2020 } from './view-configs/satija.js';
 import { justHiglass } from './view-configs/rao.js';
 import { scAtacSeq10xPbmc } from './view-configs/tenx.js';
-import { blin2019 } from './view-configs/blin.js';
+import { blin2019, multipleOmeZarrViaRasterJson } from './view-configs/blin.js';
 import { omeNgffLegacy } from './view-configs/ome-ngff-legacy.js';
 import { hubmapIntestineSnAtacSeq } from './view-configs/hubmap.js';
 import {
@@ -62,6 +62,7 @@ export const configs = {
   'sn-atac-seq-hubmap-2020': hubmapIntestineSnAtacSeq,
   'sc-atac-seq-10x-genomics-pbmc': scAtacSeq10xPbmc,
   'blin-2019': blin2019,
+  'ome-ngff-multi': multipleOmeZarrViaRasterJson,
   'ome-ngff-v0.1': omeNgffLegacy,
   'rgb-ome-tiff': rgbOmeTiff,
   'segmentations-ome-tiff': segmentationsOmeTiff,