+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `useInitialCoordination` hook to get the values of the coordination space from the initial config, which can be used for viewState reset buttons.
 - Use `config` object reference as hook dependency when no `config.uid` is present (to support both controlled and un-controlled component cases).
 - Initialize Zustand store using closure over `createViewConfigStore` function, rather than via `useEffect`.
+- Implement basic solution for multi `ome-zarr` images via `image.raster.json`
 
 ### Changed
 - Fix Material UI import statement.